requirements.txt
python-dotenv==1.1.1
supabase==2.20.0
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==10.4.0
PyMuPDF==1.24.9
langchain==0.3.27
Whoosh==2.7.4